platform: msm8916: Dynamically map continuous splash memory region
Some devices use a different memory region for the continuous splash
feature. In this case, the correct address will be read from the
MDP registers, but it will lead to a data abort at runtime because
this memory region is not mapped in the MMU.

Dynamically update the MMU page table in case the address differs
from the default to avoid this.